brand is working. Still some minor changes needed
