+++ docs/search.json
@@ -242,7 +242,7 @@
     "href": "instructions/index.html",
     "title": "Instructional pages",
     "section": "",
-    "text": "How to search\n\n\n\n\n\nInstructional video on how to search for exercises in the itembank and upload them to TestVision\n\n\n\n\n\nOct 7, 2020\n\n\nSharon Klinkenberg\n\n\n\n\n\n\n\n\n\n\n\n\nTestVision\n\n\n\n\n\nInstructional video on how to upload a sharestats question to TestVision\n\n\n\n\n\nMay 9, 2025\n\n\nSharon Klinkenberg\n\n\n\n\n\n\nNo matching items"
+    "text": "How to search\n\n\nInstructional video on how to search for exercises in the itembank and upload them to TestVision\n\n\n\n\n\nOct 7, 2020\n\n\nSharon Klinkenberg\n\n\n\n\n\n\n\n\n\n\n\n\nTestVision\n\n\nInstructional video on how to upload a sharestats question to TestVision\n\n\n\n\n\nMay 9, 2025\n\n\nSharon Klinkenberg\n\n\n\n\n\nNo matching items"
   },
   {
     "objectID": "development/qualityProcedure.html#checklist-part-1-person-a",
@@ -277,14 +277,14 @@
     "href": "news/index.html",
     "title": "ShareStats Newsletters",
     "section": "",
-    "text": "ShareStats started!\n\n\n\n\n\nNieuwsbrief 1\n\n\n\n\n\nOct 7, 2020\n\n\nMarleen de Moor\n\n\n\n\n\n\nNo matching items"
+    "text": "ShareStats started!\n\n\nNieuwsbrief 1\n\n\n\n\n\nOct 7, 2020\n\n\nMarleen de Moor\n\n\n\n\n\nNo matching items"
   },
   {
     "objectID": "team/team.html",
     "href": "team/team.html",
     "title": "Meet the Team",
     "section": "",
-    "text": "Dimitris Pavlopoulos\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nKirsten Namesnik-Silvester\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nLidia Arends\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nMarike Polak\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nMarleen de Moor\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nNiels Smits\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nOliver Lindemann\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nSharon Klinkenberg\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nTasos Psychogyiopoulos\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nNo matching items"
+    "text": "Dimitris Pavlopoulos\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nKirsten Namesnik-Silvester\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nLidia Arends\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nMarike Polak\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nMarleen de Moor\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nNiels Smits\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nOliver Lindemann\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nSharon Klinkenberg\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nTasos Psychogyiopoulos\n\n\n\n\n\n\n\n\n\n\n\n\n\nNo matching items"
   },
   {
     "objectID": "team/people/kirsten.html",
@@ -319,7 +319,7 @@
     "href": "taxonomy/taxonomy_tree.html",
     "title": "Statistics Taxonomy",
     "section": "",
-    "text": "Below you can find the statistics taxonomy used in ShareStats. You can click on each taxonomy level te get a sub selection of only those items in the itembank."
+    "text": "Below you can find the statistics taxonomy used in ShareStats. You can click on each taxonomy level to get a sub selection of only those items in the itembank."
   },
   {
     "objectID": "taxonomy/taxonomy_tree.html#taxonomy-for-item-creation",
